requirements.txt
Flask==3.1.2
Flask-Limiter==3.12
requests==2.32.5
gunicorn==23.0.0
python-dotenv==1.1.1
pywidevine==1.8.0
beautifulsoup4==4.13.5
lxml==6.0.1
